fix(landing): 리뷰 코멘트대로 재반영 (리뷰 #5 #6 #9 #10)
- #5 탭 borderRadius 를 50% 로 변경
- #6 수학 데모의 입출력부를 MathTransPanel 클라이언트 컴포넌트로 분리,
  MathTrans 는 헤딩만 남기고 use client 제거
- #9 #10 MathTransInput 의 flex=1 두 개 제거, 좌우 폭 분배는 호출부
  (MathTransPanel 의 행)에서 selectors 로 지정
